modules/root_cause_analysis/canvas_cause_map_agent/functions/rca_canvas_builder/requirements.txt
cognite-sdk>=7.64.11
pandas
pydantic
python-dotenv
cognite-pygen